pyproject.toml
[tool.poetry]
name = "jumparound"
version = "1.0.0"
homepage = "https://github.com/M-Porter/jumparound"
description = "Quickly jump around between your projects."
authors = ["Matt Porter <mtp5129@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.3"
textual = "^0.1.12"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.9b0"
isort = "^5.10.1"
beautifulsoup4 = "^4.10.0"
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.poetry.scripts]
jumparound = "jumparound.cli:cli"
